% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/github_sourcer.R
\name{checkGithubToken}
\alias{checkGithubToken}
\title{checkGithubToken}
\usage{
checkGithubToken()
}
\value{
Returns CIRepo object if Github PAT set, otherwise returns an error.
}
\description{
Checks if Github PAT is set.
}
\examples{
# ci.repo<- checkGithubToken() # Returns CIRepo object if Github PAT set.
}

