% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/github_sourcer.R
\docType{class}
\name{CIRepo-class}
\alias{CIRepo}
\alias{CIRepo-class}
\title{CIRepo}
\format{An R6 class object.}
\usage{
CIRepo
}
\description{
A class that lets you source directly from the CI github repo rather than
download and source() or parse()
}
\section{Methods}{

\itemize{
 \item \code{sourceRCode(repo_path): Downloads R code from github, then
 sources it.}
 \item \code{parseRCode(repo_path): Like sourceRCode, but parse instead.}
}
}
\keyword{internal}

