% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S3Handler.R
\docType{class}
\name{S3Handler-class}
\alias{S3Handler}
\alias{S3Handler-class}
\title{S3Handler class and handler objects.}
\format{An R6 class object.}
\usage{
S3Handler
}
\description{
This handler object allows users to operate with S3 at a very basic level.
It is entirely a wrapper for AWS CLI.
}
\section{Methods}{

\itemize{
 \item \code{writeCredentials(aws_access_key_id, aws_secret_access_key,
                              json_file, profile_name):
   There are three potential modes of use with this method:
   1) The user provides AWS access key id and secret access key
   2) The user provides a JSON file with the credentials
   3) The user provides a profile name that already exists in the default
   credentials file (~/.aws/credentials)

   Then, the method operates appropriately and initializes an object that can
   interact with AWS
 }
 \item \code{writeS3(bucket, source_path, target_path, overwrite = FALSE):
   This method allows the user to write to S3, provided a bucket, source file
   path and target path in the bucket,
   along with whether or not the user wishes to overwrite the data.}
 \item \code{listS3(bucket, path = ""):
   Given a bucket and path (optional), the user gets an entire, recursive
   listing of files and folders from the
   combined path of the bucket and path.}
 \item \code{checkFilesS3(bucket, path):
   Given a bucket and path, the user can use this method to check whether a
   file already exists in the S3 bucket.}
}
}
\keyword{internal}

