% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/github_sourcer.R
\name{setGithubToken}
\alias{setGithubToken}
\title{setGithubToken}
\usage{
setGithubToken(pat)
}
\arguments{
\item{pat}{Github PAT available in Github user settings}
}
\value{
Nothing
}
\description{
Sets Github PAT so user can use the CIRepo object.
}
\examples{
# setGithubToken("MY_GITHUB_PAT")
}

