% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables-GDrive.R
\name{vital_signs_tables}
\alias{vital_signs_tables}
\title{vital_signs_tables}
\usage{
vital_signs_tables(client_id, client_secret, creds_file, scopes, service = F)
}
\arguments{
\item{client_id}{A Google App Engine client ID}

\item{client_secret}{A Google App Engine client secret}

\item{creds_file}{A JSON file with Google App Engine credentials in it
(preferred)}

\item{scopes}{A set of API scopes for Google Drive}

\item{service}{A boolean of whether or not the credentials are for a service
account}
}
\value{
Returns a VitalSignsData class object
}
\description{
Wrapper function for VitalSignsData
}
\examples{
# vstables <- vital_signs_tables(creds_file = "credentials.json")

# vstables <- vital_signs_tables(creds_file = "service_credentials.json",
#                                service = T)

# vstables <- vital_signs_tables(client_id = "blablabla",
#                                client_secret = "blublublu")
}

