% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables-GDrive.R
\docType{class}
\name{VitalSignsData-class}
\alias{VitalSignsData}
\alias{VitalSignsData-class}
\title{Class used to warehouse tables. Uses VitalSignsTable to handle tables.}
\format{An R6 class object.}
\usage{
VitalSignsData
}
\description{
Class used to warehouse tables. Uses VitalSignsTable to handle tables.
}
\section{Methods}{

\itemize{
 \item \code{checkToken(): Checks whether oauth token is still valid.}
 \item \code{tableListMaker(datatables, tableType, tableStore): Deprecated
 function meant to serve getTables.}
 \item \code{getTables(tableType, initialFolderID, tableSave): Retrieves
 tables and stores them in a list as VitalSignsTable objects.}
 \item \code{getInternalData(): Loads only data from the Internal Data
 Folder.}
 \item \code{getExternalData(): Loads only data from the External Data
 Folder.}
 \item \code{showTables(): Shows the names of loaded tables.}
 \item \code{saveData = function(newData,
                                 newTitle,
                                 dataType,
                                 parent_folder = self$vsoutputfolderid,
                                 file_loc = NULL): Saves data based on
 the parameters provided by the user.}
}
}
\keyword{internal}

