% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables-GDrive.R
\docType{class}
\name{VitalSignsTable-class}
\alias{VitalSignsTable}
\alias{VitalSignsTable-class}
\title{VitalSignsTable}
\format{An R6 class object.}
\usage{
VitalSignsTable
}
\description{
Class used to manage individual tables.
}
\section{Methods}{

\itemize{
 \item \code{getData(returnData): Loads the data from Google Drive.
 Optionally returns the data, but defaults to not returning it.}
}
}
\keyword{internal}

